moving trapFocus to mhv/shared
